+++ odoo/local-src/qoqa_claim_unclaimed_process/__openerp__.py
@@ -1,12 +1,13 @@
 # -*- coding: utf-8 -*-
-# © 2016 Camptocamp SA (Matthieu Dietrich)
+# © 2016-2017 Camptocamp SA (Matthieu Dietrich)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
 {
     'name': 'QoQa Claim Unclaimed Process',
-    'version': '9.0.1.0.0',
+    'version': '9.0.1.0.1',
     'category': 'Others',
     'depends': [
         'crm_claim_mail',
+        'crm_rma_by_shop',
         # TODO import connector_qoqa when pay_by_email API is available
         # 'connector_qoqa'
     ],